+++ Regionalied FeliX/Data Processing Scripts/population/population_aggregation.py
@@ -2,7 +2,7 @@
 """
 Created: Thur 08 Feburary 2024
 Description: Scripts to aggregate population data by region from UNPD to 5 FeliX regions
-Scope: FeliX model regionalization, module Population 
+Scope: FeliX model regionalization, module Population
 Author: Quanliang Ye
 Institution: Radboud University
 Email: quanliang.ye@ru.nl
@@ -15,11 +15,14 @@ from pathlib import Path
 
 import numpy as np
 import pandas as pd
-import pyam
+
+# import pyam
 import yaml
 
 timestamp = datetime.datetime.now()
 file_timestamp = timestamp.ctime()
+
+project_name = "felix_regionalization"
 
 # predefine the data source of population
 # data sources are wittgensteinwittgenstein, world_bank, unpd
@@ -34,8 +37,8 @@ with open(yaml_dir, "r") as dimension_file:
 version = data_info["version"]
 felix_module = data_info["module"]
 # Any path consists of at least a root path, a version path, a module path
-path_clean_data_folder = Path(data_info["data_output"]["root_path"]).joinpath(
-    f"version_{version}/{felix_module}"
+path_clean_data_folder = Path(data_info["data_root_path"]).joinpath(
+    f"clean_data/{project_name}/{version}/{felix_module}"
 )
 
 # set logger
@@ -69,22 +72,27 @@ logging.info("Information of input data is loaded")
 
 logging.info("Set path to the raw data")
 # Any path consists of at least a root path, a version path, a module path
-path_raw_data_folder = Path(raw_data_info["root_path"]).joinpath(
-    f"version_{version}/{felix_module}"
+path_raw_data_folder = Path(data_info["data_root_path"]).joinpath(
+    f"raw_data/{project_name}/{version}/{felix_module}"
 )
 raw_data_files = raw_data_info["data_file"]
 logging.info("Path to raw data set")
 
 logging.info("Set concordance tables of regional classifications")
 # set paths of concordance table
-path_concordance_folder = Path(raw_data_info["root_path"]).joinpath(
-    f"version_{version}/concordance"
+path_concordance_folder = Path(data_info["data_root_path"]).joinpath(
+    f"raw_data/{project_name}/{version}/concordance"
 )
 concordance_file = raw_data_info["concordance"]
 logging.info("Concordance tables of regional classifications set")
 
 logging.info("Extracting dimension information for data cleaning and restructing")
-regions = data_info["dimension"]["region"]
+if "ipcc_r6" in concordance_file:
+    regions = data_info["dimension"]["ipcc_r6"]
+    final_region_name = "ipcc_r6"
+else:
+    regions = data_info["dimension"]["region"]
+    final_region_name = "un_regions"
 genders = data_info["dimension"]["gender"]
 age_cohorts = data_info["dimension"]["age"]
 logging.info("Extracted dimensions of regions, genders, and ages")
@@ -131,7 +139,7 @@ concordance_table = pd.read_csv(
     encoding="utf-8",
 )
 concordance_table = concordance_table.dropna()
-concordance_table["un_region_code"] = concordance_table["un_region_code"].astype("int")
+# concordance_table["un_region_code"] = concordance_table["un_region_code"].astype("int")
 logging.info(f"Finish reading concordance table")
 
 
@@ -228,10 +236,15 @@ def data_cleaning_unpd(
     -------
     cleaned data in pd.Dataframe
     """
+    if "ipcc_r6" in list(concordance.columns):
+        final_region_name = "ipcc_r6"
+    else:
+        final_region_name = "un_region"
+
     if download_method == "api":
         raw_data_merge = pd.merge(
             raw_data,
-            concordance[["region_api", "un_region"]],
+            concordance[["region_api", final_region_name]],
             left_on="Location",
             right_on="region_api",
         ).rename(
@@ -245,13 +258,13 @@ def data_cleaning_unpd(
     else:
         raw_data_merge = pd.merge(
             raw_data,
-            concordance[["region", "un_region"]],
+            concordance[["region", final_region_name]],
             left_on="Region, subregion, country or area *",
             right_on="region",
         )
 
     years = np.unique(raw_data_merge["Year"])
-    raw_data_groups = raw_data_merge.groupby(["un_region", "sex", "Year"])
+    raw_data_groups = raw_data_merge.groupby([final_region_name, "sex", "Year"])
     cleaned_data = []
     for region in regions:
         for sex in genders:
@@ -273,7 +286,7 @@ def data_cleaning_unpd(
                     for age in age_cohorts:
                         age = age.replace("--", "-")
                         entry = {
-                            "un_region": region,
+                            final_region_name: region,
                             "sex": sex,
                             "age": age,
                             "year": year,
@@ -286,18 +299,20 @@ def data_cleaning_unpd(
                         cleaned_data.append(entry)
                         del entry, age
                 else:
-                    entry = {
-                        "un_region": region,
-                        "sex": sex,
-                        "age": age,
-                        "year": year,
-                        "value": raw_data_region_year[age].sum()
-                        * 1000,  # convert to person
-                        "unit": "person",
-                    }
+                    for age in age_cohorts:
+                        age = age.replace("--", "-")
+                        entry = {
+                            final_region_name: region,
+                            "sex": sex,
+                            "age": age,
+                            "year": year,
+                            "value": raw_data_region_year[age].sum()
+                            * 1000,  # convert to person
+                            "unit": "person",
+                        }
 
-                    cleaned_data.append(entry)
-                    del entry, age
+                        cleaned_data.append(entry)
+                        del entry, age
 
     cleaned_data = pd.DataFrame(cleaned_data)
     cleaned_data = cleaned_data.astype({"year": "int"})
@@ -380,20 +395,20 @@ def data_restructure(
     To restructure data into the format:
       '''
           Time,1950,1951,1952,...
-          Population[Africa],x,x,x,...
-          Population[AsiaPacific],x,x,x,...
+          Population[Region 1],x,x,x,...
+          Population[Region 2],x,x,x,...
           ......
-          Population[WestEU],x,x,x,...
+          Population[Region N],x,x,x,...
           Population[World],x,x,x,...
-          Population by Gender[Africa,female],x,x,x,...
-          Population by Gender[Africa,male],x,x,x,...
+          Population by Gender[Region 1,female],x,x,x,...
+          Population by Gender[Region 1,male],x,x,x,...
           ......
-          Population by Gender[WestEU,female],x,x,x,...
-          Population by Gender[WestEU,male],x,x,x,...
+          Population by Gender[Region N,female],x,x,x,...
+          Population by Gender[Region N,male],x,x,x,...
           Population by Gender[World,female],x,x,x,...
           Population by Gender[World,male],x,x,x,...
-          Population Cohorts[Africa,female,"0-4"],x,x,x,...
-          Population Cohorts[Africa,female,"5-9"],x,x,x,x,...
+          Population Cohorts[Region 1,female,"0-4"],x,x,x,...
+          Population Cohorts[Region 1,female,"5-9"],x,x,x,x,...
           ......
           Population Cohorts[World,male,"95-99"],x,x,x,...
           Population Cohorts[World,male,"100+"],x,x,x,x,...
@@ -419,7 +434,14 @@ def data_restructure(
 
     restructured_data = []
     logging.info("Sum up cleaned data by gender and age cohort")
-    cleaned_data_by_region = cleaned_data.groupby(["un_region", "year"])["value"].sum()
+    if "ipcc_r6" in cleaned_data.columns:
+        final_region_name = "ipcc_r6"
+    else:
+        final_region_name = "un_region"
+
+    cleaned_data_by_region = cleaned_data.groupby([final_region_name, "year"])[
+        "value"
+    ].sum()
     for region in regions:
         entry = {"parameter": f"Population[{region}]"}
         for year in range(1900, 2101):
@@ -437,7 +459,7 @@ def data_restructure(
     del cleaned_data_by_region
 
     logging.info("Sum up cleaned data by age cohort")
-    cleaned_data_no_age = cleaned_data.groupby(["un_region", "sex", "year"])[
+    cleaned_data_no_age = cleaned_data.groupby([final_region_name, "sex", "year"])[
         "value"
     ].sum()
     for region in regions:
@@ -459,7 +481,7 @@ def data_restructure(
     del cleaned_data_no_age
 
     logging.info("Group cleaned data by all dimensions")
-    cleaned_data_groups = cleaned_data.groupby(["un_region", "sex", "age"])
+    cleaned_data_groups = cleaned_data.groupby([final_region_name, "sex", "age"])
     for region in regions:
         for sex in genders:
             for age in age_cohorts:
@@ -508,44 +530,44 @@ cleaned_population = data_cleaning(
 logging.info(f"Finish cleaning the raw data from {data_source}")
 
 ##################################################################################################
-logging.info(f"Add SSP2 population data from IIASA based on the specified concordance")
-logging.info(f"Load IIASA data")
-iiasa_data_file = "1706548837040-ssp_basic_drivers_release_3.0_full.csv"
-iiasa_data = pyam.IamDataFrame(
-    data=Path("data_regionalization_raw")
-    / f"version_{version}"
-    / "iiasa"
-    / iiasa_data_file
-)
-logging.info(f"Select population-related data from IIASA dataset")
-iiasa_raw_data = (
-    iiasa_data.filter(scenario="SSP2", variable="Population*", level=2)
-    .timeseries()
-    .reset_index()
-)
+# logging.info(f"Add SSP2 population data from IIASA based on the specified concordance")
+# logging.info(f"Load IIASA data")
+# iiasa_data_file = "1706548837040-ssp_basic_drivers_release_3.0_full.csv"
+# iiasa_data = pyam.IamDataFrame(
+#     data=Path("data_regionalization_raw")
+#     / f"version_{version}"
+#     / "iiasa"
+#     / iiasa_data_file
+# )
+# logging.info(f"Select population-related data from IIASA dataset")
+# iiasa_raw_data = (
+#     iiasa_data.filter(scenario="SSP2", variable="Population*", level=2)
+#     .timeseries()
+#     .reset_index()
+# )
 
-logging.info(f"Load concordance between IIASA regions and FeliX regions")
-concordance_file_iiasa = "iiasa_countries_to_5_un_regions.csv"
-concordance_table_iiasa = pd.read_csv(
-    path_concordance_folder / concordance_file_iiasa,
-    encoding="utf-8",
-)
-concordance_table_iiasa = concordance_table_iiasa.dropna().reset_index(drop=True)
-concordance_table_iiasa["un_region_code"] = concordance_table_iiasa[
-    "un_region_code"
-].astype("int")
-logging.info(f"Finish reading concordance table for IIASA")
+# logging.info(f"Load concordance between IIASA regions and FeliX regions")
+# concordance_file_iiasa = "iiasa_countries_to_5_un_regions.csv"
+# concordance_table_iiasa = pd.read_csv(
+#     path_concordance_folder / concordance_file_iiasa,
+#     encoding="utf-8",
+# )
+# concordance_table_iiasa = concordance_table_iiasa.dropna().reset_index(drop=True)
+# concordance_table_iiasa["un_region_code"] = concordance_table_iiasa[
+#     "un_region_code"
+# ].astype("int")
+# logging.info(f"Finish reading concordance table for IIASA")
 
-cleaned_data_iiasa = data_cleaning_iiasa(
-    raw_data=iiasa_raw_data,
-    concordance=concordance_table_iiasa,
-)
+# cleaned_data_iiasa = data_cleaning_iiasa(
+#     raw_data=iiasa_raw_data,
+#     concordance=concordance_table_iiasa,
+# )
 ##################################################################################################
 
 
 # Start restructuring cleaned data, which will be used as historic data
 logging.info("Start restructing the cleaned data based on the specified concordance")
-including_iiasa = True
+including_iiasa = False
 if including_iiasa:
     restructured_population = data_restructure(
         cleaned_data=pd.concat(
@@ -575,7 +597,9 @@ else:
 
     logging.info(f"Start writing the restructured data")
     restructured_population.to_csv(
-        path_clean_data_folder.joinpath(f"population_by_time_series_{data_source}.csv"),
+        path_clean_data_folder.joinpath(
+            f"population_by_time_series_{data_source}_{final_region_name}.csv"
+        ),
         encoding="utf-8",
         index=False,
     )